nyc-real-estate-ai/requirements.txt
pandas==2.0.3
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.28.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1